Refactor main.py
Made a main_loop.py file that can be imported to other files, like webapp.py. main.py now only handles logic for when the program is run as a script. So webapp.py, for example, calls main_loop.main_loop() instead of main.main().
